requirements.txt
Click>=8.0.1
pyyaml>=6.0
